refactor: replace double with BigDecimal for all monetary values
Replace all double fields and method signatures with BigDecimal across
models (Position, SandboxUser, TradeRecord, LimitOrder, StopOrder,
PriceAlert), services (SandboxTradingService, CurrencyStatTask,
SharesStatTask) and the command handler (MessageHandler). Use
BigDecimal.valueOf() for API conversions, .compareTo() for comparisons,
.divide(divisor, 8, HALF_UP) for division, and .setScale(2, HALF_UP)
for display formatting. ConfigLoader sandbox methods now return
BigDecimal. All 28 tests updated and passing.

+++ src/main/java/services/sandbox/model/LimitOrder.java
@@ -1,6 +1,7 @@
 package services.sandbox.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.Instant;
 
 /**
@@ -15,7 +16,7 @@ public class LimitOrder implements Serializable {
     private String side;
     private int qty;
     /** Target price at which the order should be executed */
-    private double limitPrice;
+    private BigDecimal limitPrice;
     private Instant createdAt;
 
     public LimitOrder() {
@@ -23,7 +24,7 @@ public class LimitOrder implements Serializable {
 
     public LimitOrder(String id, String userId, String userName,
                       String ticker, String side, int qty,
-                      double limitPrice, Instant createdAt) {
+                      BigDecimal limitPrice, Instant createdAt) {
         this.id = id;
         this.userId = userId;
         this.userName = userName;
@@ -40,6 +41,6 @@ public class LimitOrder implements Serializable {
     public String getTicker() { return ticker; }
     public String getSide() { return side; }
     public int getQty() { return qty; }
-    public double getLimitPrice() { return limitPrice; }
+    public BigDecimal getLimitPrice() { return limitPrice; }
     public Instant getCreatedAt() { return createdAt; }
 }
