+++ src/main/java/services/sandbox/model/Position.java
@@ -1,22 +1,19 @@
-/*
- * Decompiled with CFR 0.152.
- */
 package services.sandbox.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
-public class Position
-implements Serializable {
+public class Position implements Serializable {
     private String userId;
     private String ticker;
     private String instrumentId;
     private int quantity;
-    private double avgPrice;
+    private BigDecimal avgPrice;
 
     public Position() {
     }
 
-    public Position(String userId, String ticker, String instrumentId, int quantity, double avgPrice) {
+    public Position(String userId, String ticker, String instrumentId, int quantity, BigDecimal avgPrice) {
         this.userId = userId;
         this.ticker = ticker;
         this.instrumentId = instrumentId;
@@ -56,12 +53,11 @@ implements Serializable {
         this.quantity = quantity;
     }
 
-    public double getAvgPrice() {
+    public BigDecimal getAvgPrice() {
         return this.avgPrice;
     }
 
-    public void setAvgPrice(double avgPrice) {
+    public void setAvgPrice(BigDecimal avgPrice) {
         this.avgPrice = avgPrice;
     }
 }
-
