+++ src/main/java/services/sandbox/model/StopOrder.java
@@ -1,6 +1,7 @@
 package services.sandbox.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.Instant;
 
 /**
@@ -12,14 +13,14 @@ public class StopOrder implements Serializable {
     private String id;
     private String userId;
     private String ticker;
-    private double triggerPrice;
+    private BigDecimal triggerPrice;
     private Instant createdAt;
 
     public StopOrder() {
     }
 
     public StopOrder(String id, String userId, String ticker,
-                     String type, double triggerPrice, Instant createdAt) {
+                     String type, BigDecimal triggerPrice, Instant createdAt) {
         this.id = id;
         this.userId = userId;
         this.ticker = ticker;
@@ -32,6 +33,6 @@ public class StopOrder implements Serializable {
     public String getUserId() { return userId; }
     public String getTicker() { return ticker; }
     public String getType() { return type; }
-    public double getTriggerPrice() { return triggerPrice; }
+    public BigDecimal getTriggerPrice() { return triggerPrice; }
     public Instant getCreatedAt() { return createdAt; }
 }
